// src/ts/Stats.d.ts
export declare class Stats {
  domElement: HTMLElement;
  update: () => void;
}